Added async by leveraging the queue logic. Separated the Linq workflow execution, versioning and stats, fixed bugs, reduced the replica from 3 to 1,
