+++ api-gateway/src/main/java/org/lite/gateway/service/LinqWorkflowService.java
@@ -1,31 +1,49 @@
 package org.lite.gateway.service;
 
-import org.lite.gateway.dto.LinqRequest;
-import org.lite.gateway.dto.LinqResponse;
-import org.lite.gateway.dto.TeamWorkflowStats;
-import org.lite.gateway.model.LinqWorkflowStats;
 import org.lite.gateway.entity.LinqWorkflow;
-import org.lite.gateway.entity.LinqWorkflowExecution;
-import org.lite.gateway.entity.LinqWorkflowVersion;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 public interface LinqWorkflowService {
-    Mono<LinqResponse> executeWorkflow(LinqRequest request);
+    /**
+     * Create a new workflow
+     * @param workflow The workflow to create
+     * @return Mono containing the created workflow
+     */
     Mono<LinqWorkflow> createWorkflow(LinqWorkflow workflow);
-    Mono<LinqWorkflow> updateWorkflow(String workflowId, LinqWorkflow updatedWorkflow);
-    Mono<Void> deleteWorkflow(String workflowId);
-    Flux<LinqWorkflow> getWorkflows();
+
+    /**
+     * Get a workflow by ID
+     * @param workflowId The ID of the workflow to retrieve
+     * @return Mono containing the workflow if found
+     */
     Mono<LinqWorkflow> getWorkflow(String workflowId);
-    Mono<LinqWorkflowExecution> trackExecution(LinqRequest request, LinqResponse response);
-    Flux<LinqWorkflowExecution> getWorkflowExecutions(String workflowId);
-    Flux<LinqWorkflowExecution> getTeamExecutions();
-    Mono<LinqWorkflowExecution> getExecution(String executionId);
-    Flux<LinqWorkflow> searchWorkflows(String searchTerm);
-    Mono<LinqWorkflowStats> getWorkflowStats(String workflowId);
-    Mono<LinqWorkflow> createNewVersion(String workflowId, LinqWorkflow updatedWorkflow);
-    Mono<LinqWorkflow> rollbackToVersion(String workflowId, String versionId);
-    Flux<LinqWorkflowVersion> getVersionHistory(String workflowId);
-    Mono<LinqWorkflowVersion> getVersion(String workflowId, String versionId);
-    Mono<TeamWorkflowStats> getTeamStats();
+
+    /**
+     * Update an existing workflow
+     * @param workflowId The ID of the workflow to update
+     * @param updatedWorkflow The updated workflow data
+     * @return Mono containing the updated workflow
+     */
+    Mono<LinqWorkflow> updateWorkflow(String workflowId, LinqWorkflow updatedWorkflow);
+
+    /**
+     * Delete a workflow
+     * @param workflowId The ID of the workflow to delete
+     * @return Mono<Void> indicating completion
+     */
+    Mono<Void> deleteWorkflow(String workflowId);
+
+    /**
+     * Get all workflows for the current team
+     * @return Flux of workflows
+     */
+    Flux<LinqWorkflow> getAllWorkflows();
+
+    /**
+     * Search workflows by name or description
+     * @param query The search query
+     * @return Flux of matching workflows
+     */
+    Flux<LinqWorkflow> searchWorkflows(String query);
 }
