+++ api-gateway/src/main/java/org/lite/gateway/service/LinqWorkflowStatsService.java
@@ -1,0 +1,20 @@
+package org.lite.gateway.service;
+
+import org.lite.gateway.dto.TeamWorkflowStats;
+import org.lite.gateway.model.LinqWorkflowStats;
+import reactor.core.publisher.Mono;
+
+public interface LinqWorkflowStatsService {
+    /**
+     * Get statistics for a workflow
+     * @param workflowId The ID of the workflow
+     * @return Mono containing workflow statistics
+     */
+    Mono<LinqWorkflowStats> getWorkflowStats(String workflowId);
+
+    /**
+     * Get team-wide workflow statistics
+     * @return Mono containing team statistics
+     */
+    Mono<TeamWorkflowStats> getTeamStats();
+}
