+++ docker-compose-ec2.yml
@@ -252,7 +252,7 @@ services:
       - ./api-gateway:/app/api-gateway
       - ./keys:/app/keys
     deploy:
-      replicas: 3
+      replicas: 1
       restart_policy:
         condition: on-failure
     expose:
